set CPU affinity
I had been having difficulties seeing the kind of behaviour described in the
book so I emailed Mr. Kerrisk about it. He (kindly) pointed out that the
reason I was not seeing the behaviour I had been expecting to see was because
my program was running on a multi-cpu system without specifying any affinity.

In this commit I set an affinity so all processes run on the same CPU, thereby
giving me the behaviour I expected to see.