doodle_static_42-1: Q&A
Ask and answer a question about static linking: does an entire module get
pulled in when only one function of that module is requested? Also, does
everything from one static library get pulled in when something is requested?

Static linking is smart enough to pull in only from the module that is being
requested. Unfortunately it pulls in everything from that module even though
only one thing is requested.

The "moral" is: when creating a static archive, try to make each module as
small as possible since everything from that module will get pulled in. But
only modules that contain things needed are pulled in.