doodle_44-1: ask parent/child pipe question
Looking through Michael's Listing 44-3 I can't help ask myself: "is it really
necessary for the child to explicitly close() the pipe when it is done,
wouldn't the pipe get closed on exit()?". This program seeks to answer that
question and the answer is that the pipe is closed on the child's _exit(),
therefore an explicitly close() isn't necessary.

It would be necessary if the child then wanted to go on doing other things and
the close() was just meant for communication at a particular time in its
execution. But for this simple case, close()ing then _exit()ing makes the
close() redundant.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>